split libcontainer-init on its own file
