+++ libcontainer_dyno_driver.go
@@ -15,7 +15,6 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
-	"runtime"
 	"strconv"
 	"strings"
 	"syscall"
@@ -367,74 +366,6 @@ func (ctx *containerInit) startCallback() {
 	//child process is already running, it's safe to close the parent's read
 	//side of the pipe
 	ctx.configPipe.Close()
-}
-
-type LibContainerInitDriver struct{}
-
-func (dd *LibContainerInitDriver) Build(*Release) error {
-	// noop
-	return nil
-}
-
-// Start acts as PID=1 inside a container spawned by libcontainer, doing the
-// required setup and re-exec'ing as the abspath driver
-// TODO: drop privileges (setuid)
-func (dd *LibContainerInitDriver) Start(ex *Executor) error {
-	configPipe := os.NewFile(4, "configPipe")
-	var container libcontainer.Config
-	decoder := gob.NewDecoder(configPipe)
-	if err := decoder.Decode(&container); err != nil {
-		configPipe.Close()
-		return err
-	}
-	configPipe.Close()
-
-	dynoEnv := make(map[string]string, len(container.Env))
-	for _, entry := range container.Env {
-		pieces := strings.SplitN(entry, "=", 2)
-		dynoEnv[pieces[0]] = pieces[1]
-	}
-	hs := Startup{
-		App: AppSerializable{
-			Version: ex.Release.version,
-			Env:     dynoEnv,
-			Slug:    ex.Release.slugURL,
-			Stack:   ex.Release.stack,
-			Processes: []FormationSerializable{
-				{
-					FArgs:     ex.Args,
-					FQuantity: 1,
-					FType:     ex.ProcessType,
-				},
-			},
-		},
-		OneShot:     true,
-		SkipBuild:   false,
-		StartNumber: ex.ProcessID,
-		Action:      Start,
-		Driver:      &AbsPathDynoDriver{},
-		FormName:    ex.ProcessType,
-		LogplexURL:  ex.logplexURLString(),
-	}
-	args := []string{"/usr/bin/setuidgid", "dyno", "/tmp/hsup"}
-	container.Env = []string{"HSUP_CONTROL_GOB=" + hs.ToBase64Gob()}
-
-	runtime.LockOSThread() // required by namespaces.Init
-
-	// TODO: clean up /tmp/hsup and /tmp/slug.tgz after abspath reads them
-	return namespaces.Init(
-		&container, container.RootFs, "",
-		os.NewFile(3, "controlPipe"), args,
-	)
-}
-
-func (dd *LibContainerInitDriver) Stop(*Executor) error {
-	panic("this should never be called")
-}
-
-func (dd *LibContainerInitDriver) Wait(*Executor) *ExitStatus {
-	// this should be unreachable, but in case it is called, sleep forever:
-	select {}
 }
 
 func containerConfig(
